Refact : 캘린더 조회 API, Request Type 변경(Int -> String)
