+++ src/main/java/homeat/backend/domain/analyze/repository/FinanceDataRepository.java
@@ -1,16 +1,10 @@
 package homeat.backend.domain.analyze.repository;
 
 import homeat.backend.domain.analyze.entity.FinanceData;
-import homeat.backend.domain.analyze.repository.querydsl.FinanceRepositoryCustom;
-import homeat.backend.domain.user.entity.Member;
+import homeat.backend.domain.analyze.repository.querydsl.FinanceDataRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Optional;
-
 @Repository
-public interface FinanceDataRepository extends JpaRepository<FinanceData, Long>, FinanceRepositoryCustom {
+public interface FinanceDataDataRepository extends JpaRepository<FinanceData, Long>, FinanceDataRepositoryCustom {
 }
